+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 All notable changes to the "split-html-attributes" extension will be documented in this file.
 
+### [1.0.2] - Nov 21 2019
+
+- Added `sortOrder` option to specify a preferred sort order for attributes using an array of regex's
+
 ### [1.0.1] - Nov 21 2019
 
-- Added option to place closing bracket (> or />) on new line
+- Added `closingBracketOnNewLine` option to place closing bracket (> or />) on new line
 
 ### [1.0.0] - Nov 19 2019